Support null-ordering
It turns out that this _is_ supported by the
`com.healthmarketscience.sqlbuilder` library after all, but not when
using the `addCustomOrdering` method. Instead you have to use
`addCustomOrderings` (plural) which takes as its arguments or or more
`OrderObject`s, each of which may have `.setNullOrder` invoked on it.

Thanks to James Ahlborn for pointing this out to me:
https://sourceforge.net/p/openhms/feature-requests/28/#9ea1